more converted for #139
